recipe_api_test/requirements.txt
django
djangorestframework
django-filter
django-toolbelt
connection